% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holcmapr_app.R
\name{run_holcmapr}
\alias{run_holcmapr}
\title{Function to run redlining mapping methodology comparison app.}
\usage{
run_holcmapr()
}
\description{
Function to run redlining mapping methodology comparison app.
}
\examples{
if (interactive()){
  run_holcmapr()
}
}
